requirements.txt
torch==1.9.0+cu102
torchvision==0.10.0+cu102
numpy==1.19.5
pandas==1.1.5
Pillow==7.1.2
tqdm==4.41.1
matplotlib==3.2.2
